% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomeworFunction.R
\name{homework_function}
\alias{homework_function}
\title{Homework}
\usage{
homework_function(Longitude, Latitude, ...)
}
\arguments{
\item{Longitude}{The coordinate longitude to transform}

\item{Latitude}{The coordinate Latitude to transform}

\item{...}{Provides opportunity use different argument to affect Longitude and Latitude}
}
\description{
Transform coordinate from decimal minutes into decimal degrees
}
\examples{
homework_function()
}
\keyword{Latitude}
\keyword{Longitude}
\keyword{Transformation}
