% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homewok2_function_edit.R
\name{homework_function}
\alias{homework_function}
\title{Homework}
\usage{
homework_function(Longitude = "Longitude", Latitude = "Latitude", mean = FALSE)
}
\arguments{
\item{Longitude}{The coordinate longitude to transform}

\item{Latitude}{The coordinate Latitude to transform}

\item{...}{Provides opportunity use different argument to affect Longitude and Latitude}
}
\description{
Transform coordinate from decimal minutes into decimal degrees
}
\examples{
homework_function <- function(Longitude = "Longitude", Latitude = "Latitude", mean = FALSE)
}
\keyword{Latitude}
\keyword{Longitude}
\keyword{Transformation}
